ADE-744 = separate lib and cli modules
