+++ cli/package.json
@@ -1,35 +1,42 @@
 {
-  "name": "lib",
+  "name": "cli",
   "version": "0.0.0",
-  "main": "dist",
   "license": "UNLICENSED",
   "private": true,
   "engines": {
     "node": ">=10"
   },
   "scripts": {
-    "build": "tsc",
+    "build": "NODE_ENV=production rollup -c --environment BUILD:production",
+    "start": "NODE_ENV=development rollup -c",
+    "watch": "npm run start -- -w",
     "lint": "eslint src --ext .ts,.tsx",
     "coverage": "type-coverage",
     "format": "prettier --write 'src/**/*.{ts,tsx}'",
-    "test": "jest"
+    "test": "echo TODO add tests"
   },
   "dependencies": {
+    "clear": "^0.1.0",
+    "clui": "^0.3.6",
     "colors": "^1.4.0",
     "date-fns": "^2.19.0",
     "fp-ts": "^2.10.4",
+    "inquirer": "^8.0.0",
     "io-ts": "^2.2.15",
     "io-ts-types": "^0.5.15",
-    "jira-client": "^6.21.1",
+    "lib": "file://../lib",
     "monocle-ts": "^2.3.9",
     "newtype-ts": "^0.3.4",
     "oauth": "^0.9.15",
     "readonly-types": "^2.0.0",
-    "ts-pattern": "^3.1.2"
+    "ts-pattern": "^3.1.2",
+    "yargs": "^16.2.0"
   },
   "devDependencies": {
+    "@types/clear": "^0.1.1",
+    "@types/clui": "^0.3.0",
+    "@types/inquirer": "^7.3.1",
     "@types/jest": "^26.0.22",
-    "@types/jira-client": "^6.21.0",
     "@types/oauth": "^0.9.1",
     "@typescript-eslint/eslint-plugin": "^4.21.0",
     "@typescript-eslint/parser": "^4.21.0",
@@ -43,8 +50,14 @@
     "eslint-plugin-total-functions": "^4.7.2",
     "jest": "^26.6.3",
     "prettier": "^2.2.1",
+    "rollup": "^2.45.2",
+    "rollup-plugin-cleanup": "^3.2.1",
+    "rollup-plugin-delete": "^2.0.0",
+    "rollup-plugin-execute": "^1.1.1",
+    "rollup-plugin-typescript2": "^0.30.0",
     "string-length": "^4.0.2",
     "ts-jest": "^26.5.5",
+    "tsconfig-paths-webpack-plugin": "^3.5.1",
     "type-coverage": "^2.17.2",
     "typescript": "^4.2.4"
   },
