+++ package.json
@@ -7,7 +7,8 @@
     "node": ">=10"
   },
   "workspaces": [
-    "lib"
+    "lib",
+    "cli"
   ],
   "scripts": {
     "build": "npm run build --workspaces",
